Remove hard dependency on proper
Running "make eunit" will now download proper into deps and ensure it
is on the code path. Projects that depend on jiffy will no longer pick
up a transitive dependency on proper.

When the eunit target is run, proper will be cloned into deps/ if not
already found and compiled (also if not already compiled).

Also includes a few Makefile tweaks:

* Add distclean target
* Add .PHONY for Makefile target hygeine