Also check whether the linker supports -flto
Currently the LTO (link time optimisation) check is done by trying
to compile a single object file with the -flto option. This way the
check does not cover the case that the compiler supports it but the
linker does not.

This can happen for instance on FreeBSD11 with FreeBSD clang version
6.0.0 in combination with GNU ld 2.17.50, where the latter does not
support the -plugin option.

Modify the check by removing the -c flag from the compiler options.